chore: move modules in respective folders
