+++ postreise/analyze/generation/summarize.py
@@ -1,0 +1,23 @@
+from powersimdata.scenario.scenario import Scenario
+from powersimdata.scenario.analyze import Analyze
+
+
+def sum_generation_by_type_zone(scenario):
+    """Sums generation for a Scenario in Analyze state by {type, zone}.
+    :param powersimdata.scenario.scenario.Scenario scenario: scenario instance.
+    :return: (*pandas.DataFrame*) -- total generation, indexed by {type, zone}.
+    :raise Exception: if scenario is not a Scenario object in Analyze state.
+    """
+    if not isinstance(scenario, Scenario):
+        raise TypeError('scenario must be a Scenario object')
+    if not isinstance(scenario.state, Analyze):
+        raise ValueError('scenario.state must be Analyze')
+
+    pg = scenario.state.get_pg()
+    grid = scenario.state.get_grid()
+    plant = grid.plant
+
+    summed_gen_series = pg.sum().groupby([plant.type, plant.zone_id]).sum()
+    summed_gen_dataframe = summed_gen_series.unstack().fillna(value=0)
+
+    return summed_gen_dataframe
