+++ postreise/analyze/generation/tests/__init__.py
@@ -1,0 +1,1 @@
+__all__ = ['test_carbon', 'test_curtailment', 'test_summarize']
