+++ postreise/analyze/generation/tests/test_curtailment.py
@@ -1,11 +1,10 @@
 import unittest
 
-import numpy as np
-from numpy.testing import assert_array_equal, assert_array_almost_equal
+from numpy.testing import assert_array_equal
 import pandas as pd
 
 from postreise.tests.mock_scenario import MockScenario
-from postreise.analyze.curtailment import \
+from postreise.analyze.generation.curtailment import \
     calculate_curtailment_time_series, calculate_curtailment_percentage, \
     summarize_curtailment_by_bus, summarize_curtailment_by_location
