+++ postreise/analyze/tests/__init__.py
@@ -1,1 +1,1 @@
-__all__ = ['test_carbon', 'test_mwmiles', 'test_statelines']
+__all__ = ['test_helpers']
